npogrpi/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==3.2.25
django-filter==23.5
django-js-asset==2.0.0
django-mptt==0.14.0
django-tinymce==3.6.1
django_debug_toolbar==3.8.1
djangorestframework==3.15.1
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.11.5
mccabe==0.7.0
Pillow==9.5.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2024.1
sorl-thumbnail==12.9.0
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
